+++ library/SolverUtils/DriverArpack.cpp
@@ -376,7 +376,7 @@ void DriverArpack::v_Execute(ostream &out)
             WriteEvs(pFile, i, dr[i], di[i]);
 
             std::string file = m_session->GetSessionName() + "_eig_" +
-                               boost::lexical_cast<std::string>(i) + ".fld";
+                               std::to_string(i) + ".fld";
             WriteFld(file, z + i * n);
         }
     }
@@ -392,7 +392,7 @@ void DriverArpack::v_Execute(ostream &out)
                      false);
 
             std::string file = m_session->GetSessionName() + "_eig_" +
-                               boost::lexical_cast<std::string>(i) + ".fld";
+                               std::to_string(i) + ".fld";
             WriteFld(file, z + i * n);
         }
     }