+++ CHANGELOG.md
@@ -56,6 +56,7 @@ v5.5.0
 - Add a particle tracker utilitiy using equation system infrastructure (!1310)
 - Remove deprecated fs::copy_directory function call (!1662)
 - Added a sample python script for wallnormaldata module (!1663)
+- Replace boost::lexical_cast with std::to_string for integer variable (!1677)
 
 v5.4.0
 ------